+++ ethereum/package.json
@@ -47,6 +47,7 @@
     "solidity-docgen": "^0.5.10",
     "temp": "^0.9.1",
     "ts-node": "^10.0.0",
+    "tsconfig-paths": "^4.1.0",
     "typechain": "^8.1.0",
     "typescript": "^4.3.2",
     "unique-filename": "^1.1.1"
