+++ ethereum/tsconfig.json
@@ -9,10 +9,15 @@
     "outDir": "dist",
     "allowSyntheticDefaultImports": true,
     "declaration": true,
-    "declarationMap": true
+    "baseUrl": ".",
+    "paths": { "@src": ["src"] },
   },
+  "ts-node": {
+    "require": ["tsconfig-paths/register"]
+  },
+  "exclude": ["node_modules", "dist"],
   "include": [
-    "./deploy", "./scripts", "./tasks", "./src",
+    "deploy", "scripts", "tasks", "src", "test"
   ],
   "files": ["./hardhat.config.ts"],
 }
